Make project to get trace.info, generated and dependencies files updated. Merge leftovers in QueriesGenerated gone with regenerated generators. Classes in new models (mps.core) got minor updates due to fixes in original branch

+++ plugins/newLanguageDesign/languages/syntax/source_gen/jetbrains/mps/core/syntax/behavior/SNonTerm_Behavior.java
@@ -20,7 +20,7 @@ public class SNonTerm_Behavior {
   }
 
   public static boolean virtual_isList_1030525575875844509(final SNode thisNode) {
-    return (int) ListSequence.fromList(SLinkOperations.getTargets(thisNode, "rules", true)).count() == 2 && (int) ListSequence.fromList(SLinkOperations.getTargets(thisNode, "rules", true)).translate(new ITranslator2<SNode, SNode>() {
+    return ListSequence.fromList(SLinkOperations.getTargets(thisNode, "rules", true)).count() == 2 && ListSequence.fromList(SLinkOperations.getTargets(thisNode, "rules", true)).translate(new ITranslator2<SNode, SNode>() {
       public Iterable<SNode> translate(SNode it) {
         return SLinkOperations.getTargets(it, "parts", true);
       }
